Refactor transifex handlers
