+++ lib/txgh/app.rb
@@ -1,9 +1,6 @@
-require 'base64'
-require 'json'
 require 'sinatra'
 require 'sinatra/json'
 require 'sinatra/reloader'
-require 'uri'
 
 module Txgh
 
@@ -24,6 +21,7 @@ module Txgh
   end
 
   class Hooks < Sinatra::Base
+    include Txgh::Handlers
     # Hooks are unprotected endpoints used for data integration between Github and
     # Transifex. They live under the /hooks namespace (see config.ru)
 
@@ -43,47 +41,21 @@ module Txgh
 
 
     post '/transifex' do
-      settings.logger.info('Processing request at /hooks/transifex')
-      settings.logger.info(request.inspect)
-
-      payload = Hash[URI.decode_www_form(request.body.read)]
-      config = Txgh::KeyManager.config_from_project(payload['project'])
-
-      if authenticated_transifex_request?(config.transifex_project, request)
-        handler = Txgh::Handlers::TransifexHookHandler.new(
-          project: config.transifex_project,
-          repo: config.github_repo,
-          resource_slug: payload['resource'],
-          language: payload['language'],
-          logger: settings.logger
-        )
-
-        handler.execute
-        status 200
-      else
-        status 401
-      end
-    end
-
-    post '/github' do
-      response = Txgh::Handlers::Github::RequestHandler.handle_request(
+      status_code, body = Transifex::RequestHandler.handle_request(
         request, settings.logger
       )
 
-      status response.status
-      json response.body
+      status status_code
+      json body
     end
 
-    private
+    post '/github' do
+      status_code, body = Github::RequestHandler.handle_request(
+        request, settings.logger
+      )
 
-    def authenticated_transifex_request?(project, request)
-      if project.webhook_protected?
-        TransifexRequestAuth.authentic_request?(
-          request, project.webhook_secret
-        )
-      else
-        true
-      end
+      status status_code
+      json body
     end
   end
