+++ lib/txgh/handlers.rb
@@ -1,7 +1,6 @@
 module Txgh
   module Handlers
-    autoload :Github,               'txgh/handlers/github'
-    autoload :TransifexHookHandler, 'txgh/handlers/transifex_hook_handler'
-    autoload :Response,             'txgh/handlers/response'
+    autoload :Github,    'txgh/handlers/github'
+    autoload :Transifex, 'txgh/handlers/transifex'
   end
 end
