+++ lib/txgh/handlers/transifex.rb
@@ -1,0 +1,8 @@
+module Txgh
+  module Handlers
+    module Transifex
+      autoload :HookHandler,    'txgh/handlers/transifex/hook_handler'
+      autoload :RequestHandler, 'txgh/handlers/transifex/request_handler'
+    end
+  end
+end
